SystemDialog: force title and subtitle labels to support Rich Text
The sefault setting of "mightBeRichText" does not behave correctly with
KUIT markup, so if either of these labels do use any KUIT markup, it is
not displayed properly. The heuristic used is fast and simple; since we
know that we want KUIT markup to be supported, let's bypass it and
manually specify that rich text is supported here.