disable automatic portal launching early on
this opts out of color picking support early on in the login process
when our environment is still incomplete. not doing this would cause the
portal system to launch "too early" and have an incomplete environment
set. instead disable the portal support early on and only enable it once
we have a working environment constructed.

also move cursor setup and ksplash into this newly available "critical
section". now that we have an opt out we can run them as early as
necessary without risking them starting the portal infrastructure.

BUG: 458865