applets/batterymonitor: make active for PPD modes while plugged in
When performance mode has been activated, sometimes it's hard to know or
remember--save for your computer's fans being loud, if they're loud.
This can waste energy and make the user feel less in control of their
computer.

And when power saver mode has been held by an app, it can make your
computer slower, so you'll want to know that too.

Accordingly, this commit makes the Battery & Brightness applet become
active (and hence visible in the system tray) while plugged in and
either Performance is active, or Power Save mode has been automatically
held by an app. Additionally, the applet's icon changes to that of the
active mode, and its tooltip tells you why.

Nothing changes if the machine is not plugged in, since in this case,
seeing the battery charge level is more important.