kcms/regionandlanguage: convert ASCII unicode symbol to QChar
BUG: 460704

The locale files in `/usr/share/i18n/locales` contain ASCII Unicode symbol, we need to replace it with real Unicode rune. 

**country_name "T\<U00FC>rkiye"** to **country_name "Türkiye"**

When we are at it, also add caching to `parseLocaleFile`, otherwise we parse multiple locale format files every time we want address, phone number examples.

And add an autotest for all of this.