Add a TriangleMouseFilter component
Extracts the triangle mouse filter from the Kicker plugin into its own
component, importable as org.kde.plasma.workspace.trianglemousefilter 1.0.

Mostly the same as the Kicker version but with new properties, to enable its
eventual use in the Task Manager:

- blockFirstEnter, default false (unlike Kicker plugin)
- edgeLine, to specify how long the target edge is
- secondaryPoint, another check which makes filtering more accurate in TM case

A wrapper is included which exports it under the old name with the old defaults.