Port libnotificationmanager from kscreen
Kscreen has multiple nested event loops. We are hitting the notification
manager constructor from QML. This could explain any number of
unexplained backtraces.

It's also needless to fetch information we already have from another
source. The only advantage of kscreen is that we include disabled
displays or for writing, but we don't need either of those.