applets/batterymonitor: don't show percentage when fully charged
When the user has opted to show the percentage on the battery icon,
continuing to show it when the battery is fully charged doesn't make
sense; the user knows that the battery is at either 100%, or the charge
threshold they previously set. So showing them the exact number is just
visual noise and can be safely omitted.