startkde/plasma-session: fix playing startup sound
Attaching the Phonon objects to the Startup object and then
immediately deleting the Startup object leads to the Phonon
objects being deleted too and no sound will play.

I have no idea how that ever worked, maybe Phonon used to
work differently in the past or maybe it depends on the backend.

Since the MediaObject is configured to autodelete itself
after it finishes playing, it doesn't need to be parented
to anything.

CCBUG: 422948