applets/batterymonitor: only show when PPD mode is held by an app
Making the icon show and use the PPD mode's icon anytime it's in
performance mode is overkill and causes visual noise when the system is
configured to *always* be in performance mode while plugged in, which is
a valid and supported configuration. In this case and the case where the
user has manually enabled performance or power save mode, noyifying is
probably not needed since the users knows and expects that the system is
using this mode.

Instead, let's only make the widget active and use the PPD mode's icon
when it's been held that way by an app, which is a thing that happens
automatically and otherwise the user might not know it's happened.