SystemDialog: Fix display when there is no component
This is something that can happen when we are relying only on the title
and the dialog buttons.
We were not taking into account the cases when it was null so it would
fail to calculate the minimum size.


(cherry picked from commit 384501fe2229453c04c5a6f8ae25dbab14dec1ac)