Only consider screens to be mirrored if they are completely overlapping
Right now libnotificationmanager considers screens to be mirrored--and
hence automatically enables Do Not Disturb Mode--when they are
overlapping at all. But this doesn't match KScreen's definition of when
screens are mirrored, which is that one screen is *completely* covered
up by the other. Let's use that same logic here so we're not enabling
DND mode when screens are simply overlapping a bit, which is a valid if
unusual configuration in KScreen.