kcms/users: Replace finger selection combobox with graphic of fingers
The names of specific fingers are one of the rare strings where it's hard, if not possible to pick a satisfactory name that works for everyone and doesn't result in the UI sounding too formal/clinical or too casual/regional; especially across languages, dialects, and varying levels of language proficiency.
The best way to bypass issues with words is to not use them at all (except for people who need a screen reader that uses words), so the combobox gets replaced with an interactive image of fingers on hands.
Also, it looks cool.