kcms/color: Fix wrong color application of --accent-color flag
The --accent-color flag was not updated to keep with the changes and bugfixes
and so it contains some already fixed bugs. It can't update the header area
when tint is used. To fix this its own way of color application has been
replaced with the better plasmashell API which will always be kept up to date.