ksplashqml: Set cursor shape per window
At the moment, the cursor is not hidden on wayland because there's an
issue in qtwayland qpa related to handling of QGuiApplication::overrideCursor()
which I plan to fix.

On the other hand, it would be nicer to use a more declarative way to
manage the cursor, i.e. set the cursor shape for each individual window
and let Qt take care of the rest rather than use ugly
QGuiApplication::setOverrideCursor().