logout-greeter: Simplify shutdown check
Shutdown functionality was moved out of ksmserver to make that more X11
only. We moved to a wrapper plasma-shutdown that has more generic calls,
including telling ksmserver to close gracefully.

org.kde.Shutdown is a shortlived program that only is alive whilst we
are trying to perform a logout. Querying the name should suffice.