applets/digitalclock: return to old tooltip code
d2b33b72d5e76d30f5fe6403468538ba3ba4cc1b replaced the old code with a
Kirigami.FormLayout, which was not correct because the labels always use
the app color scheme, not the Plasma color scheme, causing unreadable
text when using a dark Plasma theme with a light app color scheme.

This commit returns to the old code and makes a tiny fix to it to make
the layout work properly (which was the reason I ported to Formlayout in
the first place, so that was a wholly unnecessary change given the
ability to fix the issue in another way).