applets/digital-clock: refine timezone config UI
Right now the ability to add your home timezone as its own list item to
make it appear automatically when traveling is a hidden feature; let's
make it obvious by adding some text to clarify things. To further
clarify the timezone-related functionalty on thie page, let's also unify
terminology and adjust the positioning of the existing message to better
connect it with the UI control it refers to.