[KSplashQML] Re-add logging "Loading stage" to debug output.
It got removed in faf76152ef because back then it did
the time measuring itself.
But it would be useful to diagnose if not all stages are received.

Removing also the unused logMessageHandler.

To use it modify ~/.config/QtProject/qtlogging.ini
[Rules]
org.kde.plasma.ksplashqml=true

Then the output should appear in `journalctl --user -f` or
for x11 still in `tail -f -n 0 ~/.xsession-errors | ts`.