wallpapers/image: ctx menu entry for setting desktop/lockscreen wallpaper
When browsing through images and finding a suitable lockscreen
background, its cumbersum to then have to navigate to it once again in
the image picker in settings.

We want to avoid increasing the amount of entries in an already large
context menu, so the lockscreen/desktop/both entries are aggregated
under a "Set as" menu entry.