Improve responsiveness of triangular filter
The responsiveness of the filter is improved by adding two new
metrics to the filter logic:     - Firstly, the direction of the cursor
movement is considered. If (in the default configuration) the filter is
currently active and the user moves the cursor to the left (away from the
content area), the filter stops intercepting the movement events. This
results in a noticeable improvement in responsiveness without compromise
the objective of the filter.     - Secondly, a velocity metric (as
already brought up in BUG 438950) is added. This ensures that a very slow
movement of the cursor results in disabling the triangular filter. This,
also, improves the the feeling of responsiveness significantly. Still,
there is a need for the timer in the filter, as when the cursor has
stopped its movement completely, it will no longer send HoverMove events
to the filter, thus it wouldn't be possible to detect a complete stop
of the cursor.

(cherry picked from commit 6c18efd13cabde92f82fa881a7ebb29beda7344b)

CCBUG: 467426