kcms/cursortheme: add explanation to launch feedback dialog and adjust label
Following Andy's suggestion:

I would like to suggest a little more helpful text for the user. Maybe at the top of the dialog add a paragraph like this:
"Select options to control the way the cursor blinks while waiting for processes to run or applications to launch."
Also, instead of "Cursor", I would suggest calling this "Feedback:" or "Cursor Feedback". This way, it's more clear that you are "only" working with the feedback while waiting.

Also, we want to use active verbs in our strings to call users to action. The language needs to read more naturally. For the task manager section, I would suggest:

[X] Enable Task Manager Animation