kcms/lnf: Simplify the representation of theme's contents
Combine the enum flags `LayoutToApply` and `AppearanceToApply`
into a single `Contents` enum, which represents all the different
contents that a Global Theme can have and the user can select.

The Manager class just uses the contents selection when saving.
To simplify the data flow and reduce a layer of complexity, store
this selection on the main KCM class.