kcms/lnf: Simplify theme contents selection pages
Some refactoring to reduce a bit of code duplication.

The checkbox for the "Include Desktop Layout" can be included
in the model of the Repeater below, without need for a special
case

Specially do not make items visibility in `SimplePage`
depend on one another, since that's usually more confusing
and error prone.