notifications: Improve activation infrastructure
The Wayland activation protocol xdg_activation_v1 requires us to send
the input serial that triggers the notification as one of the inputs. So
far we were always using 0 because it's not trivial to infer which
window is the one that triggers the action.

This change adds API to specify the calling window so this is not a
problem anymore.

BUG: 464099