Fix path skipping screen removal signal
We might get screen removal signals and then a re-ordering from the
backend or the other way round.

In the event that we get the re-order first it's the intention to take
this as the single source of truth and remove it from
m_availableScreens. In this case we have to notify clients. It's
important that this removal is signalled before we process a re-order.

BUG:464171