screenpool: Update screenInvariants check to match code changes
On wayland it's possible to get the output order for a screen removal
before the output deletion signal. Screen pool is designed to handle
this and the current code is designed so that output order takes
precendence and will remove the screen from availableScreens
immediately.

This is fine in real life but one of our asserts needs guarding.