Make clocks tooltip consistent and add seconds
Having the seconds visible in your clock widget can be distracting, so most people decide to keep them hidden. Sometimes though it happens that you may need to know them: the idea  behind this mr is to make them always available in the tooltip of the clock widgets. And also make these tooltips consistent with each others as a bonus.

Unfortunately I'm no expert with qml, and I got stuck trying to accessing the tooltip inside the widget qml sources. I need to know when the tooltip is shown in order to change the sampling time and the alignment type of the time datasource. In the case of the digital clock, I can get that through `Plasmoid.toolTipItem.visible` because a custom tooltip is used, but in the analog clock this isn't the case and so `Plasmoid.toolTipItem` doesn't reference anything. I'd need to access the `visible` property of the `ToolTipDialog` currently in use, and probably a good solution would be to connect that to a new signal in `appletinterface`, but I haven't even understood how that is, for example, connected to the `ToolTip` class for example, so any help/suggestion would be welcomed.
Otherwise I could make a custom tooltip in the analog clock too and use the same trick as for the digital one, but it sounds more like a hack than a solution.

PS: As soon as I fix the problem above I'll take care of the fuzzy and binary clocks too.

BUG: 463853
FIXED-IN: 6