X11: Do not set the GDK_SCALE / GDK_DPI_SCALE environment variables
The primary method for setting the global and text scale for GTK applications is by using the `gdk-window-scaling-factor` and `gdk-unscaled-dpi` settings. Those are also read by GTK4, which at the same time dropped usage of GDK_DPI_SCALE.

Note that the scaling settings are only relevant on X11; On Wayland the DPI scale is communicated directly by KWin to applications on a per-monitor basis.

See also https://invent.kde.org/plasma/kde-gtk-config/-/merge_requests/49

CCBUG: 442901
FIXED-IN: 5.27