StatusNotifierWatcher: Always claim there is a StatusNotifierHost
This is of course conceptually wrong. But toolkits may have to
provide a static check if the system tray is available or
to switch the used backend  at startup. This includes for example
Qt (when not using our platform theme) or Electron.
Pragmatically it makes no difference on Plasma as any Xembed
fallback would also not show up in a non-existent system tray.
The applications just make one DBus call more if the setup of the
user doesn't include any system tray.
BUG:425315
FIXED-IN:5.27