Revert "startplasma-wayland: Don't set GDK_SCALE and GDK_DPI_SCALE"
This reverts commit 63de3d6f1716a6924fa5924b5a57af4abaa4460f.

This change was incorrect and the original logic for it was backwards;
it was stated that keeping it causes non-GTK XWayland (e.g. Electron)
apps to be scaled incorrectly. But in fact the opposite was true;
getting rid of it was what causes them to be scaled incorrectly! This
mistake was made due to due to the presence of a loca workaround on my
system that inverted the behavior.

BUG: 465733
FIXED-IN: 5.27.1

CCBUG: 443215


(cherry picked from commit 34e3efd85252182977178ab99435b99a1ba987b9)