applets/digital-clock: fix logic error in tooltip
The clocks model adds two items for each timezone, so if we want to
check for when there's only a single timezone configured, we need to
check if the model has 3 or more items in it, not 2.

That was fixed in a521f5ada3dadbe938c662a607aa28737530bd8e for master
but not packported to Plasma 5.27. This commit does so.

BUG: 465873
FIXED-IN: 5.27.1