applets/digitalclock: restore smaller tooltip for simple case
When using the "Show seconds: never" setting, there's no need to show the
time and time zone in the tooltip for the simple case where there's only
one timezone; the information is redundant.

This commit fixes that by returning to the simple two-line tooltip
content for that combination of settings.