shell: Remove X11 clear color workaround in DesktopView
Ideally DesktopView should be opaque as there's nothing to paint below
it. We tried to do this a couple of times in the past, but it resulted
in the panel losing its translucency on X11.

The main culprit is that QtQuick could reuse the same OpenGL context
for every window, which might be undesired because the windows may have
mismatching EGLConfigs, etc.

With the introduction of RHI, it's no longer the case. So let's remove
the X11 workaround and see if it finally works as expected.