Screenpool: avoid uniqueConnection with lambda
Handling QScreen geometryChanged was connected to a lambda with a unique
connection. Qt cannot determine this and doesn't connect.

The unique connection was used as handleScreenAdded is also called when
the output order changed to handle things being async.
(I don't think that's actually needed, but that's a refactor for another thread)

This patch shuffles the code so we only connect once which is cleaner
anyway.

CCBUG: 466312
CCBUG: 466020