klipper: remove duplicate items when loading from history
Due to a catastrophic bug existing in 5.27 beta and 5.27.0, there could
be duplicate items in klipper's history file, so filter them out here.

The popup menu will try to read `item->next_uuid()` in a loop to fill the menu, hence it's possible to create an infinite loop when there are duplicate items.

BUG: 466236
CCBUG: 465225
FIXED-IN: 5.27.4