Don't show unnecessary auth dialog when LUKS password is found in wallet
Currently when mounting a LUKS-encrypted volume, if a saved password is
found in KWallet, the auth dialog still appears, but it has the password
field pre-populated. This is a bit odd and doesn't match what typically
happens when there's a saved password: it's simply sent along, without
showing a dialog at all. This commit does that here too for a smoother
workflow.

BUG: 467098
FIXED-IN: 6.0