kcms/nightcolor: fix timing strings with narrow window widths
These labels lived in an item that had `anchors.centerIn: parent` set on
it, which makes it have no maximum width and it can overflow. Clearly it
was the intent to avoid that, because all the child items had their
properties set up correctly for it.

To fix this, instead anchor to the parent's size and center the
contents, and then make the parent fill the available width.

Resolves https://invent.kde.org/plasma/plasma-mobile/-/issues/244