ksmserver: help users with intentionally read-only homedirs self-satisfy
As part of its sanity check, ksmserver doesn't start Plasma if `$HOME`
is read-only unless `KDE_HOME_READONLY=1` is set in the environment. But
this is never told to the user, so an advanced user or system
administrator who has intentionally set up the system this way won't
know what to do about it and will submit a bug report to KDE.

To resolve this, inform such users about the relevant environment
variable in the error message, without giving specific details about how
to set it. Advanced users will already know how to do this and thus
self-satisfy.

BUG: 463352
FIXED-IN: 6.0