Introduce libclock
Designed as a simpler replacement to the datetime engine, with some key
differences to the internals.

- Tracking of the system timezone is handled internally and mostly
abstracted, the caller gets information about the effective timezone
regardless of whether it's the system or defined. This means we don't
need to worry about using UTC and adding offsets in QML code.

- We use absolute timers throughout; we're using timerfd for clock skew
notifications anyway, we may as well use this to align timers on the
minute.

Backend timers are shared, front-end Clock classes are designed to be
thin wrappers.