applets/kicker: Replace RunnerMatchesModel sorting with KRunner::ResultsModel
We still need to provide additional features, like the context menu actions.
But the sorting will be the same.

Also, each KRunner::ResultsModel takes care of managing the RunnerManager internally.
Meaning we can not have one shared RunnerManager and then redistribute the results
to different models in case we have mergeResults=false.
Instead, we have multiple ResultsModels in case mergeResults is false, but
each RunnerManager loads exactly one runner.

This also means that we should not delete the models after the query changed,
because the plugin instantiation and initialization would cause overhead.

BUG: 431204
FIXED-IN: 6.0