applets/mediacontroller: Change song title opacity on play/pause
When media is playing, the song title has the normal opacity of 1,
but when paused it will change to 0.6, matching the artist. This
provides feedback on whether media is playing, and the change is
animated.

---

An example use case for this is when a user might set their
headphones down and is unsure of whether media is playing. With
this, they can see at a glance and middle click to pause.