applets/digital-clock: reload timezone after saving in Datetime KCM
The datetime KCM emits a signal `clockUpdated` after the timezone is updated. Listen to the signal and reload the model when the signal is emitted. After the model is updated, re-apply selected timezones or select the updated default timezone when there is no selected timezone in the config dialog.

This allows the default timezone in the widget's config dialog to be updated immediately after system timezone changes.

BUG: 467494
FIXED-IN: 5.27.4