Introduce porting aid xdg-shell integration plugin fork
In order to port plasma to the layer-shell protocol, plasma must be able
to create a popup with a layer surface as the parent.

With Qt 6.5 and older, there's no nicer way other than provide a custom
shell integration plugin. In Qt 6.6, we introduced an API that
layer-shell-qt could use to associate xdg popups with it.

Qt 6.6 will be released in September if everything goes smooth, but it's
not ideal to wait until that point to start the porting process.

This change introduces a temporary fork of the xdg-shell integration
plugin to allow us start the porting process now. When Qt 6.6 is released,
the fork shell integration plugin will be dropped.