Stop setting XCURSOR_THEME and XCURSOR_SIZE
Cursor size is specified in different coordinate spaces on Wayland and
X11. Wayland uses logical pixels, while X11 uses native pixels.

Due to X11 using native pixels, we need to provide an upscaled cursor
size in Wayland session to match the cursor size in Wayland apps.

That's not possible while plasma sets the XCURSOR_SIZE environment
variable, which applies both to X11 and Wayland apps.

Besides that, the environment variable is not a good way to propagate
the settings in general as it's not possible to get notified when the
environment variables change.