Don't always honor natural scrolling setting for widget icons
In both the Battery and Media widgets, there is a clear directionality
to the thing being changed by a scroll; you're turning the volume or
brightness UP or DOWN. Even when using the inverted scrolling setting,
you want for a scroll up to raise the brightness, and vice versa.

Same diff as https://invent.kde.org/plasma/plasma-pa/-/merge_requests/171
but for the Battery & Brightness widget.