runners/calculator: set timeout to prevent allocating huge memory
This is not exactly a memory leak. The extreme expression simply
requires huge amounts of memory to calculate.

The calculator runner is problematic since it does not specify
a time limit for calculations, and does not even abort
calculations after the window has been closed
(a simple example of this is the expression 123456789!).

Fixes https://github.com/Qalculate/libqalculate/issues/527

BUG: 468084
FIXED-IN: 5.27.5