Make "Shut down" the default logout option
When pressing Ctrl+Alt+Del, the power button, or one of the remaining
"Leave" actions, you'll see the logout screen with the default logout
action selected. Currently that default action is "Log Out."

Most people use single-user machines for which "Shut down" is probably a
better default setting. For this reason, switch to Shut Down as the
default action in the logout screen when invoking it from Ctrl+Alt+Del,
the power button, or one of the remaining "Leave" actions.

BUG: 467532
FIXED-IN: 6.0