klipper: Expose DBus signal to tell when clipboardHistory was updated
The signal can be used to watch for changes on the clipboard for
example to wait for the users input on other program or interoperate
in any other way.

An example use case is if a user want to sync the kill-ring of Emacs
or any other Editor with Klipper.

BUG:424685
FIXED-IN: 6.0

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>