outputorderwatcher: ignore outputs without crtc
when turning outputs **off** using `xrandr --off` that output is still
connected but won't have any crtc attached. this is distinctly different
from when kscreen disables an output, which indeed marks the output
disconnected but it still has a crtc

previously we would expect to find screens without crtc in the
list of `qApp->screens()` which would then trigger an infinite refresh
loop via the refresh timer because the screen would never be `present`
on account of being off.

BUG: 466362